Create Wordle.cpp
Wordle

Chef invented a modified wordle.

There is a hidden word S and a guess word T, both of length 5.

Chef defines a string M to determine the correctness of the guess word. For the ith index:

If the guess at the ith index is correct, the ith character of M is G.
If the guess at the Ith index is wrong, the ith character of M is B.
Given the hidden word S and guess T, determine string M.